Add new remote control service
This patch exposes the new screenshot functionality to applications.

An application which is signed by an authorised key can use the new
RemoteControlService to grab screen images and to inject input events.
Signing keys can be authorised at OS build time or via a new internal
API with "signature" level permission. Alternatively remote control
can be enabled globally; the user will have to dismiss a warning
message to enable this.

The service is accessed via the android.os.RemoteControl Java package.
API documentation is provided in this package in Javadoc format, and
ends up in the standard SDK documentation.