Command-line tool to request a screenshot from the SurfaceFlinger
This tool uses the new screen grabbing API to request an image of the
screen, and writes it to standard output in the format expected by the
DDMS debugging tool.

This is intended to be used as a more reliable alternative to reading
directly from the frame buffer. A separate patch gives adbd the
ability to run this tool when the debugger asks to capture the screen.

Change-Id: I0ebeeb247f84c23db4260287f63ddb9a1c22c529